Moving SenderService and SenderWorker to org.acra.sender package. Moving CrashReport* classes to org.acra.common package.
