+++ src/main/java/org/acra/sender/ReportSender.java
@@ -34,7 +34,7 @@ public interface ReportSender {
     /**
      * Send crash report data.
      *
-     * Method will be called from the {@link org.acra.SenderService}.
+     * Method will be called from the {@link SenderService}.
      *
      * @param context       Android Context in which to send the crash report.
      * @param errorContent  Stores key/value pairs for each report field.
